+++ sentry/src/main/java/io/sentry/SpanDataConvention.java
@@ -26,4 +26,8 @@ public interface SpanDataConvention {
   String HTTP_START_TIMESTAMP = "http.start_timestamp";
   String HTTP_END_TIMESTAMP = "http.end_timestamp";
   String PROFILER_ID = "profiler_id";
+  String CACHE_HIT = "cache.hit";
+  String CACHE_KEY = "cache.key";
+  String CACHE_OPERATION = "cache.operation";
+  String CACHE_WRITE = "cache.write";
 }